fix codacy issues

+++ UltraStar Play/Assets/Scenes/SongEditor/OverviewBar/IssueOverviewVisualizer.cs
@@ -73,7 +73,7 @@ public class IssueOverviewVisualizer : MonoBehaviour, INeedInjection, ISceneInje
 
         // Use a minimum width of 0.5% such that the issue is not overlooked
         int lengthInMillis = endMillis - startMillis;
-        double lengthInPercent = lengthInMillis / songDurationInMillis;
+        double lengthInPercent = (double)lengthInMillis / songDurationInMillis;
         if (lengthInPercent < 0.005)
         {
             endMillis = (int)(startMillis + songDurationInMillis * 0.005);
